Added implementation of getRoom and roomExists
